requirements.txt
Django==1.8.5
djangorestframework==3.3.0
futures==3.0.3
